api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
flask = "^3.1.0"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"

[tool.poetry.scripts]
start = "run:start_app"
db-upgrade = "api.scripts:upgrade_db"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
